r10079: Fix DevID RevID and VendorID to be 32 bit values. The wrapper was auto generated after that fix

+++ ibdm/datamodel/Fabric.h
@@ -246,9 +246,9 @@ class IBNode {
  public:
   string			   name;         // Name of the node (instance name of the chip)
   IBNodeType 		type;      // Either a CA or SW
-  uint16_t        devId;     // The device ID of the node
-  uint16_t        revId;     // The device revision Id.
-  uint16_t        vendId;    // The device Vendor ID.
+  uint32_t        devId;     // The device ID of the node
+  uint32_t        revId;     // The device revision Id.
+  uint32_t        vendId;    // The device Vendor ID.
   uint8_t         rank;      // The tree level the node is at 0 is root.
   class IBSystem *p_system;  // What system we belong to
   class IBFabric *p_fabric;  // What fabric we belong to.
